fix: remove asterisk from filename
